Refactor code
- Put all in a single file: `src/index.js`
- Remove all breaking changes
