--- src/styles/basics.js
@@ -1,14 +1,0 @@
-import { style, compose } from '../style'
-
-export const opacity = style({
-  prop: 'opacity',
-})
-
-export const overflow = style({
-  prop: 'overflow',
-})
-
-export const basics = compose(
-  opacity,
-  overflow,
-)
