--- src/variant.js
@@ -1,4 +1,0 @@
-import { get } from './util'
-
-export const variant = ({ key, prop = 'variant' }) => props =>
-  get(props.theme, [key, props[prop]].join('.')) || null
