--- src/variants.js
@@ -1,5 +1,0 @@
-import { variant } from './variant'
-
-export const buttonStyle = variant({ key: 'buttons' })
-export const textStyle = variant({ key: 'textStyles', prop: 'textStyle' })
-export const colorStyle = variant({ key: 'colorStyles', prop: 'colors' })
