+++ test/styles.js
@@ -35,7 +35,6 @@ test('returns raw color values', t => {
 })
 
 // Impossible to ensure, due to perf issues
-
 // test('backgroundColor prop overrides bg prop', t => {
 //   const a = color({
 //     backgroundColor: 'tomato',
@@ -44,16 +43,14 @@ test('returns raw color values', t => {
 //   t.deepEqual(a, [{ backgroundColor: 'tomato' }])
 // })
 
-// Useless, font-size default to pixels
-
-// test('returns a pixel font-size', t => {
-//   const a = fontSize({ fontSize: 48 })
-//   t.deepEqual(a, { fontSize: '48px' })
-// })
+test('returns a pixel font-size', t => {
+  const a = fontSize({ fontSize: 48 })
+  t.deepEqual(a, { fontSize: '48px' })
+})
 
 test('uses a default font-size scale', t => {
   const a = fontSize({ fontSize: 2 })
-  t.deepEqual(a, { fontSize: 16 })
+  t.deepEqual(a, { fontSize: '16px' })
 })
 
 test('returns a string font-size', t => {
@@ -90,7 +87,7 @@ test('returns 0 values', t => {
 
 test('returns negative pixel values', t => {
   const styles = space({ m: -2 })
-  t.deepEqual(styles, { margin: -8 })
+  t.deepEqual(styles, { margin: '-8px' })
 })
 
 test('returns negative em values', t => {
@@ -105,7 +102,7 @@ test('returns negative theme values', t => {
     },
     m: -2,
   })
-  t.deepEqual(styles, { margin: -8 })
+  t.deepEqual(styles, { margin: '-8px' })
 })
 
 test('returns positive theme values', t => {
@@ -124,8 +121,8 @@ test('returns responsive values', t => {
   })
   t.deepEqual(styles, {
     margin: 0,
-    '@media screen and (min-width: 40em)': { margin: 8 },
-    '@media screen and (min-width: 52em)': { margin: 16 },
+    '@media screen and (min-width: 40em)': { margin: '8px' },
+    '@media screen and (min-width: 52em)': { margin: '16px' },
   })
 })
 
@@ -133,7 +130,7 @@ test('returns aliased values', t => {
   const styles = space({
     px: 2,
   })
-  t.deepEqual(styles, { paddingLeft: 8, paddingRight: 8 })
+  t.deepEqual(styles, { paddingLeft: '8px', paddingRight: '8px' })
 })
 
 test('returns string values from theme', t => {
@@ -164,12 +161,12 @@ test('returns values from theme object', t => {
     margin: 'sm',
   })
 
-  t.deepEqual(styles, { margin: 1 })
+  t.deepEqual(styles, { margin: '1px' })
 })
 
 test('pl prop sets paddingLeft', t => {
   const styles = space({ pl: 2 })
-  t.deepEqual(styles, { paddingLeft: 8 })
+  t.deepEqual(styles, { paddingLeft: '8px' })
 })
 
 test('pl prop sets paddingLeft 0', t => {
@@ -177,57 +174,56 @@ test('pl prop sets paddingLeft 0', t => {
   t.deepEqual(styles, { paddingLeft: 0 })
 })
 
-// The order of props matter
+// The order of props matters
 test('px prop overrides pl prop', t => {
   const styles = space({
     pl: 1,
     px: 2,
   })
-  t.deepEqual(styles, { paddingLeft: 8, paddingRight: 8 })
+  t.deepEqual(styles, { paddingLeft: '8px', paddingRight: '8px' })
 })
 
-// The order of props matter
+// The order of props matters
 test('py prop overrides pb prop', t => {
   const styles = space({
     pb: 1,
     py: 2,
   })
-  t.deepEqual(styles, { paddingTop: 8, paddingBottom: 8 })
+  t.deepEqual(styles, { paddingTop: '8px', paddingBottom: '8px' })
 })
 
-// The order of props matter
+// The order of props matters
 test('mx prop overrides mr prop', t => {
   const styles = space({
     mr: 1,
     mx: 2,
   })
-  t.deepEqual(styles, { marginLeft: 8, marginRight: 8 })
+  t.deepEqual(styles, { marginLeft: '8px', marginRight: '8px' })
 })
 
-// The order of props matter
+// The order of props matters
 test('my prop overrides mt prop', t => {
   const styles = space({
     mt: 1,
     my: 2,
   })
-  t.deepEqual(styles, { marginTop: 8, marginBottom: 8 })
+  t.deepEqual(styles, { marginTop: '8px', marginBottom: '8px' })
 })
 
-// The order of props matter
+// The order of props matters
 test('margin overrides m prop', t => {
   const styles = space({
     m: 1,
     margin: 2,
   })
-  t.deepEqual(styles, { margin: 8 })
+  t.deepEqual(styles, { margin: '8px' })
 })
 
-// PropTypes are no longer included
-// test('space includes propTypes', t => {
-//   const { propTypes } = space
-//   t.is(typeof propTypes, 'object')
-//   t.is(typeof propTypes.m, 'function')
-// })
+test('space includes propTypes', t => {
+  const { propTypes } = space
+  t.is(typeof propTypes, 'object')
+  t.is(typeof propTypes.m, 'function')
+})
 
 test('size returns width and height', t => {
   const styles = size({
@@ -236,7 +232,7 @@ test('size returns width and height', t => {
   t.deepEqual(styles, { width: '4px', height: '4px' })
 })
 
-// // grid
+// grid
 test('gridGap returns a scalar style', t => {
   const a = gridGap({
     theme: {
@@ -245,7 +241,7 @@ test('gridGap returns a scalar style', t => {
     gridGap: 3,
   })
 
-  t.deepEqual(a, { gridGap: 8 })
+  t.deepEqual(a, { gridGap: '8px' })
 })
 
 test('gridGap uses the default scale', t => {
@@ -254,7 +250,7 @@ test('gridGap uses the default scale', t => {
     gridGap: 2,
   })
 
-  t.deepEqual(a, { gridGap: 8 })
+  t.deepEqual(a, { gridGap: '8px' })
 })
 
 test('gridRowGap returns a scalar style', t => {
@@ -265,7 +261,7 @@ test('gridRowGap returns a scalar style', t => {
     gridRowGap: 3,
   })
 
-  t.deepEqual(a, { gridRowGap: 8 })
+  t.deepEqual(a, { gridRowGap: '8px' })
 })
 
 test('gridRowGap uses the default scale', t => {
@@ -274,7 +270,7 @@ test('gridRowGap uses the default scale', t => {
     gridRowGap: 2,
   })
 
-  t.deepEqual(a, { gridRowGap: 8 })
+  t.deepEqual(a, { gridRowGap: '8px' })
 })
 
 test('gridColumnGap returns a scalar style', t => {
@@ -285,7 +281,7 @@ test('gridColumnGap returns a scalar style', t => {
     gridColumnGap: 3,
   })
 
-  t.deepEqual(a, { gridColumnGap: 8 })
+  t.deepEqual(a, { gridColumnGap: '8px' })
 })
 
 test('gridColumnGap uses the default scale', t => {
@@ -294,7 +290,7 @@ test('gridColumnGap uses the default scale', t => {
     gridColumnGap: 2,
   })
 
-  t.deepEqual(a, { gridColumnGap: 8 })
+  t.deepEqual(a, { gridColumnGap: '8px' })
 })
 
 test('textStyle prop returns theme.textStyles object', t => {
